Improve download reliability and ffmpeg packaging

+++ DownKyi.Core/BiliApi/Models/Json/Subtitle.cs
@@ -4,8 +4,8 @@ namespace DownKyi.Core.BiliApi.Models.Json;
 
 public class Subtitle : BaseModel
 {
-    [JsonProperty("from")] public float From { get; set; }
-    [JsonProperty("to")] public float To { get; set; }
+    [JsonProperty("from")] public decimal From { get; set; }
+    [JsonProperty("to")] public decimal To { get; set; }
     [JsonProperty("location")] public int Location { get; set; }
     [JsonProperty("content")] public string Content { get; set; } = string.Empty;
 }
